Add comprehensive Claude.ai POST request logging
- Create /api/root-debug endpoint to capture Claude.ai connection attempts
- Remove conflicting /app/route.ts to avoid Next.js routing conflict with page.tsx
- Enhanced logging shows User-Agent, headers, request body, and redirects
- Ready to debug INVALID_REQUEST_METHOD errors from Claude-User agent
